Tweak auth context for authenticated proxy
The previous auth context used a token in localstorage. When IAM Zero is
deployed behind an authenticated reverse proxy the proxy handles auth
and we don't need to store this token in the frontend.
When IAM Zero is run locally for demonstration purposes we can bypass
auth for the web console.