Allow configuring SQS to bypass token auth
The primary goal of the IAM Zero client tokens is to prevent
unauthenticated users from sending messages to an IAM Zero collector.

Using the SQS transport means that you need to authenticate via AWS IAM
anyway, so in some situations it may be simpler to configure IAM Zero to
not check the token and instead rely on AWS IAM policies.