Simplify recommendations
Remove "recommendations-as-code" for now and drop the Advice interface
in favour of a concrete struct.

The initial complex recommendation we were testing was creating KMS
grants for a DynamoDB table encrypted with CMK. We can likely achieve
this in a more declarative way, rather than creating a custom function
for it.